Address PR comments and refactor acceptance tests
