+++ analytics_dashboard/templates/base_dashboard.html
@@ -46,12 +46,6 @@
             {% block summary-cards %}
             {% endblock %}
 
-            <div class="view-intro row">
-              <p class="lead text-muted intro-text">
-                {% block intro-text %}{% endblock intro-text %}
-              </p>
-            </div>
-
             {% block content %}
             {% endblock content %}
           </main>
